whatsapp-support-backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
python-dotenv==1.0.*
motor==3.6.*
pydantic==2.8.*
pydantic-settings==2.4.*
twilio==9.2.*
